+++ php/CHANGELOG.md
@@ -5,8 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-* Removed license symlinks.
+## [0.1.2] - 2022-12-16
+
+### Changed
+
+- Improves error handling when parsing and invalid signature.
+- Removed license symlinks.
 
 ## [0.1.1] - 2022-08-17