+++ rust/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.1.5
+* Improves error handling when parsing and invalid signature.
+
+## 0.1.5
 * When verifying permit signed/verified path single trailing slash mismatches.
 
 ## 0.1.4